refactor the account adapter to a modular fasion
