+++ src/infrastructure/adapters/__init__.py
@@ -1,7 +1,7 @@
 """Infrastructure adapters for external systems."""
 
-from .ib_adapter import IbAdapter
-from .futu_adapter import FutuAdapter
+from .ib import IbAdapter
+from .futu import FutuAdapter
 from .file_loader import FileLoader
 from .broker_manager import BrokerManager
 from .market_data_manager import MarketDataManager
