+++ src/infrastructure/adapters/ib/__init__.py
@@ -1,0 +1,9 @@
+"""
+Interactive Brokers adapter package.
+
+Provides IbAdapter for connecting to IB TWS/Gateway.
+"""
+
+from .adapter import IbAdapter
+
+__all__ = ["IbAdapter"]
